change printed representation, add :INCLUDE option
Algebraic data types are now printed in a readable fashion, provided
the slots themselves are readable. It is recommended that users use
this notation when it's necessary, e.g., in quoted lists.

Also added is an :INCLUDE option, which does more-or-less the same
thing as the option to DEFSTRUCT, except that it only works with other
ADTs defined.

Bumped the version to 1.2.0.