basic do-notation.
it's super helpful to have do notation simplify and organize code.

it works just like #'map version, where you can specify the
result-type, but we give a bind function that will be used.

if we get to the point where we can specify all the typeclasses,
maybe we can remove the bind function argument.